[change] 別クラスに切り出すリファクタリングを行いました。
